Splitted test into multiple files
